+++ rust/adventofcode/src/intcode/io.rs
@@ -1,16 +1,16 @@
 pub trait ProgramIO {
-    fn get_next_input(&mut self) -> i32;
-    fn receive_output(&mut self, output: i32);
+    fn get_next_input(&mut self) -> i64;
+    fn receive_output(&mut self, output: i64);
 }
 
 pub struct BasicProgramIO {
-    inputs: Vec<i32>,
+    inputs: Vec<i64>,
     current_input: usize,
-    outputs: Vec<i32>,
+    outputs: Vec<i64>,
 }
 
 impl BasicProgramIO {
-    pub fn new(inputs: &[i32]) -> BasicProgramIO {
+    pub fn new(inputs: &[i64]) -> BasicProgramIO {
         BasicProgramIO {
             inputs: inputs.iter().copied().collect(),
             current_input: 0,
@@ -18,18 +18,18 @@ impl BasicProgramIO {
         }
     }
 
-    pub fn outputs(&self) -> &[i32] {
+    pub fn outputs(&self) -> &[i64] {
         &self.outputs
     }
 }
 
 impl ProgramIO for BasicProgramIO {
-    fn get_next_input(&mut self) -> i32 {
+    fn get_next_input(&mut self) -> i64 {
         self.current_input += 1;
         self.inputs[self.current_input - 1]
     }
 
-    fn receive_output(&mut self, output: i32) {
+    fn receive_output(&mut self, output: i64) {
         self.outputs.push(output);
     }
 }
